+++ src/main/java/baseball/domain/ball/BallStatus.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain.ball;
 
 public enum BallStatus {
     STRIKE, BALL, NOTHING;
