+++ src/test/java/baseball/domain/ball/BallTest.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain.ball;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
